# pantry-pal/backend/openAPI.yaml
openapi: 3.0.0
info:
  title: Pantry Management API
  description: API to manage pantry items and recipes 
  version: 1.0.0
paths:
...
  /api/recipes/{recipeId}:
    patch:
      summary: Update a favorite recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavRecipe'
      responses:
        '200':
          description: Recipe updated
    delete:
      summary: Remove a favorite recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Recipe removed
...
components:
  schemas:
    Ingredient:
      type: object
      properties:
        quantity:
          type: number
          nullable: true
        units:
          type: string
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        expirationDate:
          type: string
          format: date
          nullable: true
        frozen:
          type: boolean
          default: false
    FavRecipe:
      type: object
      properties:
        name:
          type: string
          nullable: true
        ingredients:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Ingredient'
        directions:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        planned:
          type: boolean
          default: false
        date:
          type: string
          format: date
          nullable: true
...